backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
google-generativeai==0.3.1
pydantic==2.5.3
python-dotenv==1.0.0
pandas==2.1.4
gunicorn==21.2.0
